[FIX] html_builder: fix formatting for spacing and border options
The value displayed in the input, saved in the inline style and applied
on the element sometimes differs for spacing and border.

Steps to reproduce the spacing issue:
- Click on a snippet and activate grid layout
- In the spacing option, enter "1.5"
- Leave the input to validate
=> Value: "1", Style: "1.5", Applied: "1.5"

Steps to reproduce the border issue:
- Click on a snippet
- In the border option, enter "1.5"
- Leave the input to validate
=> Input: "1", Style: "1.5", Applied: "1"

This commit fixes the issue by adding a props for inputs that should
be integer only. It also fixes the parsing issue for the spacing option,
introduce with the new builder.

task-6460556

+++ addons/html_builder/static/src/core/core_builder_action_plugin.js
@@ -283,7 +283,7 @@ export class StyleAction extends BuilderAction {
             }
             return value;
         } else if (styleName === "row-gap" || styleName === "column-gap") {
-            return parseInt(getStyleValue(el, styleName)) || 0;
+            return parseFloat(getStyleValue(el, styleName)) || 0;
         } else if (styleName === "width") {
             return el.style.width;
         } else if (styleName === "background-color") {
